Added min/max feature for temp. Working on making methods reusuable

+++ app/models/Reading.java
@@ -3,13 +3,14 @@ import javax.persistence.Entity;
 import play.db.jpa.Model;
 import utilities.Conversion;
 
+import models.Station;
 @Entity
 public class Reading extends Model {
   public int code;
   public double temperature;
   public double windSpeed;
   public int windDirection;
-  public int pressure;
+  public double pressure;
 
   public Reading(int code, double temperature, double windSpeed, int windDirection, int pressure) {
     this.code = code;
@@ -46,7 +47,7 @@ public class Reading extends Model {
   public int getWindDirection() {
     return windDirection;
   }
-  public int getPressure(){
+  public double getPressure(){
     return pressure;
   }
 
@@ -69,4 +70,5 @@ public class Reading extends Model {
     String tempIcon = Conversion.temperatureToIcon(temperature);
     return tempIcon;
   }
+
 }
